\name{BFF-package}
\alias{BFF-package}
\alias{BFF}
\docType{package}
\title{
\packageTitle{BFF}
}
\description{
\packageDescription{BFF}
}
\details{

  \tabular{ll}{
    Package: \tab BFF\cr
    Type: \tab Package\cr
    Version: \tab 2.5.0\cr
    Date: \tab 10-18-2023\cr
    License: \tab GPL>=2\cr
  }
}
\author{
\packageAuthor{BFF}

Maintainer: \packageMaintainer{BFF}
}
