% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFF.R
\name{BFF_gaussian_predictive}
\alias{BFF_gaussian_predictive}
\title{BFF Change Point Detector using Predictive Posterior P-Values for Gaussian Data}
\usage{
BFF_gaussian_predictive(
  data,
  threshold_val = c(0.05),
  burnin = 200,
  grace_period = 20,
  alpha = 39,
  beta = 1.8,
  sw = 2000,
  collect_p_val = TRUE
)
}
\arguments{
\item{data}{data stream to perform change point detection on}

\item{threshold_val}{vector of thresholds to use to assess whether new data point is a change point}

\item{burnin}{the intial period to use to build the model on which no change points are detected}

\item{grace_period}{period after a change in which changes are not detected}

\item{alpha}{alpha parameter for lambda beta prior}

\item{beta}{beta parameter for lambda beta prior}

\item{sw}{sliding window size for p-value calibration}

\item{collect_p_val}{whether or not to store p-values, boolean}
}
\value{
return the change points detected by the algorithm for each of the dection procedures, for the different thresholds. If collecting p-values then results will contain these.
}
\description{
Detects change points within data stream sequentially using a Bayesian adaptive estimation
procedure and predictive posterior p-values.
}
