% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFF_Sequential.R
\name{BFF_gaussian_seq_update}
\alias{BFF_gaussian_seq_update}
\title{BFF Change Point Detector Sequential Update}
\usage{
BFF_gaussian_seq_update(
  data,
  i,
  threshold_val = c(0.05),
  grace_period = 20,
  sw = 2000,
  alpha = 39,
  beta = 1.8,
  post_p = TRUE,
  pred_p = TRUE,
  current_mu,
  current_sigma2,
  N_values_old_old,
  D_values_old_old,
  M_values_old_old,
  N_values_old,
  D_values_old,
  M_values_old,
  mu_0_prior_value_old,
  sigma2_0_prior_value_old,
  sigma2_alpha_0_prior_value_old,
  sigma2_beta_0_prior_value_old,
  mu_0_prior_value,
  sigma2_alpha_0_prior_value,
  sigma2_beta_0_prior_value,
  anomalous_mu_threshold_sw,
  anomalous_lambda_threshold_sw,
  p_values_change_mu,
  p_values_change_lambda,
  anomalous_predictive_threshold_cal,
  p_values_predictive
)
}
\arguments{
\item{data}{new data point}

\item{i}{data point number}

\item{threshold_val}{vector of thresholds to use to assess whether new data point is a change point}

\item{grace_period}{period after a change in which changes are not detected}

\item{sw}{sliding window size for p-value calibration}

\item{alpha}{alpha parameter for lambda beta prior}

\item{beta}{beta parameter for lambda beta prior}

\item{post_p}{to calculate predictive posterior p-values}

\item{current_mu}{estimated mu from previous time}

\item{current_sigma2}{estimated sigma2 from previous time}

\item{N_values_old_old}{old old N value}

\item{D_values_old_old}{old old D value}

\item{M_values_old_old}{old old M value}

\item{N_values_old}{old N value}

\item{D_values_old}{old D value}

\item{M_values_old}{old M value}

\item{mu_0_prior_value_old}{old prior mean for mu}

\item{sigma2_0_prior_value_old}{old prior var for mu}

\item{sigma2_alpha_0_prior_value_old}{old prior alpha for sigma2}

\item{sigma2_beta_0_prior_value_old}{old prior beta for sigma2}

\item{mu_0_prior_value}{prior mean for mu}

\item{sigma2_alpha_0_prior_value}{prior alpha for sigma2}

\item{sigma2_beta_0_prior_value}{prior beta for sigma2}

\item{anomalous_mu_threshold_sw}{current anomalous mu posterior p-values list}

\item{anomalous_lambda_threshold_sw}{current anomalous lambda posterior p-values list}

\item{p_values_change_mu}{uncalibrated mu posterior p-values}

\item{p_values_change_lambda}{uncalibrated lambda posterior p-values}

\item{anomalous_predictive_threshold_cal}{current anomalous predictive posterior p-values list}

\item{p_values_predictive}{uncalibrated predictive posterior p-values}
}
\value{
return the change points detected by the algorithm for each of the dection procedures, for the different thresholds.
}
\description{
Detects change points within data stream sequentially using a Bayesian adaptive estimation
procedure and posterior p-values.
}
