% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFF.R
\name{BFF_poisson_predictive_posterior_paramest}
\alias{BFF_poisson_predictive_posterior_paramest}
\title{BFF Change Point Detector and Parameter Estimator using Predictive Posterior P-Values and Posterior P-Values for Poisson Data}
\usage{
BFF_poisson_predictive_posterior_paramest(
  data,
  threshold_val = 0.005,
  burnin = 200,
  grace_period = 20,
  alpha = 39,
  beta = 1.8,
  sw = 2000,
  param_est = FALSE
)
}
\arguments{
\item{data}{data stream to perform change point detection on}

\item{threshold_val}{vector of thresholds to use to assess whether new data point is a change point}

\item{burnin}{the intial period to use to build the model on which no change points are detected}

\item{grace_period}{period after a change in which changes are not detected}

\item{alpha}{alpha parameter for lambda beta prior}

\item{beta}{beta parameter for lambda beta prior}

\item{sw}{sliding window size for p-value calibration}

\item{param_est}{whether to estimate paramters}
}
\value{
return the change points detected by the algorithm for each of the dection procedures, for the different thresholds. Collects p-values and model parameter estimates.
}
\description{
Detects change points within data stream sequentially using a Bayesian adaptive estimation
procedure and predictive posterior p-values. Additionally estimates parameters
}
