% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background_functions.R
\name{BFF_predictive_p_value_calculator_approx}
\alias{BFF_predictive_p_value_calculator_approx}
\title{BFF predictive p value calculator}
\usage{
BFF_predictive_p_value_calculator_approx(
  x_new,
  muhat,
  N_value,
  D_value,
  M_value,
  mu_0_prior_value,
  sigma2_0_prior_value,
  sigma2_alpha_0_prior_value,
  sigma2_beta_0_prior_value
)
}
\arguments{
\item{x_new}{new data point}

\item{muhat}{mu estimate}

\item{N_value}{N Value}

\item{D_value}{D Value}

\item{M_value}{M Value}

\item{mu_0_prior_value}{mean prior value for mu}

\item{sigma2_0_prior_value}{var prior value for mu}

\item{sigma2_alpha_0_prior_value}{alpha for sigma2 prior}

\item{sigma2_beta_0_prior_value}{beta for sigma2 prior}
}
\value{
predictive posterior p-value
}
\description{
BFF predictive p value calculator
}
