% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINAL_FUNCTION_correlation_test.R
\name{cor_test_BFF}
\alias{cor_test_BFF}
\title{cor_test_BFF}
\usage{
cor_test_BFF(
  z_stat,
  n = NULL,
  alternative = "two.sided",
  r = NULL,
  omega = NULL
)
}
\arguments{
\item{z_stat}{Z statistic}

\item{n}{sample size (if one sample test)}

\item{r}{r value}

\item{tau2}{tau2 values (can be a single entry or a vector of values)}
}
\value{
Returns an S3 object of class `BFF` (see `BFF.object` for details).
}
\description{
cor_test_BFF constructs BFFs based on the z test. BFFs depend on hyperparameters r and tau^2 which determine the shape and scale of the prior distributions which define the alternative hypotheses.
By setting r > 1, we use higher-order moments for replicated studies. Fractional moments are set with r > 1 and r not an integer.
All results are on the log scale.
}
\examples{
corrBFF = cor_test_BFF(z_stat = 2.5, n = 50)
cor_test_BFF(z_stat = 2.5, n = 50, omega = 0.5)
cor_test_BFF(z_stat = 2.5, n = 50, omega = c(0.5, 0.2))
cor_test_BFF(z_stat = 2.5, n1 = 50, n2 = 40, one_sample = FALSE)
cor_test_BFF(z_stat = 2.5, n = 50, r = 2)
cor_test_BFF(z_stat = 2.5, r = 2, n1 = 50, n2 = 30, one_sample = FALSE)
cor_test_BFF(z_stat = 2.5, n = 50, r = 2.5)
cor_test_BFF(z_stat=2.5, r = 2.5, n1 = 50, n2 = 30,  one_sample = FALSE)
cor_test_BFF(z_stat = 2.5, n = 50)
cor_test_BFF(z_stat = 2.5, n = 50, omega = 0.5)
cor_test_BFF(z_stat = 2.5, n = 50, tau2 = c(0.5, 0.8))
corrBFF$BFF_max_RMSE   # maximum BFF omega
corrBFF$max_RMSE       # effect size which maximizes the BFF value

}
