% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINAL_FUNCTIONS_f_test.R
\name{f_test_BFF}
\alias{f_test_BFF}
\title{f_test_BFF}
\usage{
f_test_BFF(
  f_stat,
  n,
  df1,
  df2,
  omega = NULL,
  omega_sequence = if (is.null(omega)) seq(0.01, 1, by = 0.01),
  r = 1,
  effect_size = NULL
)
}
\arguments{
\item{f_stat}{F statistic}

\item{n}{sample size}

\item{df1}{numerator degrees of freedom.}

\item{df2}{denominator degrees of freedom.}

\item{omega}{standardized effect size on the package's internal RMSES scale (can be a single entry or a vector of values)}

\item{omega_sequence}{sequence of standardized effect sizes. If no omega is provided, omega_sequence is set to be seq(0.01, 1, by = 0.01)}

\item{r}{variable controlling dispersion of non-local priors. Default is 1. r must be >= 1}

\item{effect_size}{scale used for \code{omega} and \code{omega_sequence}. Defaults to the package's internal \code{"omega"} RMSES scale. Alternatives include \code{"cohens_f"}, \code{"cohens_f2"}, \code{"partial_eta2"}, and \code{"partial_r2"}.}
}
\value{
Returns an S3 object of class `BFF` (see `BFF.object` for details).
}
\description{
f_test_BFF constructs BFFs based on the F test. BFFs depend on hyperparameters r and tau^2 which determine the shape and scale of the prior distributions which define the alternative hypotheses.
By setting r > 1, we use higher-order moments for replicated studies. Fractional moments are set with r > 1 and r not an integer.
All results are on the log scale.
}
\examples{
fBFF = f_test_BFF(f_stat = 1.5, n = 50, df1 = 25, df2 = 48)
fBFF
plot(fBFF)

}
