% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background_functions.R
\name{gaussian_unknown_mean_var_BFF_sequential}
\alias{gaussian_unknown_mean_var_BFF_sequential}
\title{Calculate Gaussian BFF update}
\usage{
gaussian_unknown_mean_var_BFF_sequential(
  data_new,
  N_prev,
  D_prev,
  M_prev,
  mu_0,
  sigma2_0,
  alpha_0,
  beta_0,
  alpha = 39,
  beta = 1.8
)
}
\arguments{
\item{data_new}{new data point}

\item{N_prev}{N previous value}

\item{D_prev}{D previous value}

\item{M_prev}{M previous value}

\item{mu_0}{prior mean for mu}

\item{sigma2_0}{prior variance value for mu}

\item{alpha_0}{alpha for sigma^2 prior}

\item{beta_0}{beta for sigma^2 prior}

\item{alpha}{alpha for lambda prior}

\item{beta}{beta for lambda prior}
}
\value{
List containing updated hyperparameters and estimates
}
\description{
Calculate Gaussian BFF update
}
