% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background_functions.R
\name{lambda_posterior_poisson_unk_rate_sequential_beta_prior_log}
\alias{lambda_posterior_poisson_unk_rate_sequential_beta_prior_log}
\title{Lambda posterior function logged}
\usage{
lambda_posterior_poisson_unk_rate_sequential_beta_prior_log(
  lambda,
  data_new,
  N_prev,
  D_prev,
  F_prev,
  alpha_0,
  beta_0,
  alpha,
  beta
)
}
\arguments{
\item{lambda}{current lambda estimate}

\item{data_new}{new data point}

\item{N_prev}{N prev}

\item{D_prev}{D prev}

\item{F_prev}{F prev}

\item{alpha_0}{rate prior alpha}

\item{beta_0}{rate prior beta}

\item{alpha}{lambda prior alpha}

\item{beta}{lambda prior beta}
}
\value{
Value of posterior
}
\description{
Lambda posterior function logged
}
