% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background_functions.R
\name{lambda_posterior_unk_mu_sigma2_sequential_2_beta_prior_log}
\alias{lambda_posterior_unk_mu_sigma2_sequential_2_beta_prior_log}
\title{Lambda posterior function logged}
\usage{
lambda_posterior_unk_mu_sigma2_sequential_2_beta_prior_log(
  lambda,
  data_new,
  N_prev,
  D_prev,
  M_prev,
  mu_0,
  sigma2_0,
  alpha_0,
  beta_0,
  alpha = 39,
  beta = 1.8
)
}
\arguments{
\item{lambda}{last lambda value}

\item{data_new}{new data point}

\item{N_prev}{N previous value}

\item{D_prev}{D previous value}

\item{M_prev}{M previous value}

\item{mu_0}{prior mean for mu}

\item{sigma2_0}{prior variance value for mu}

\item{alpha_0}{alpha for sigma^2 prior}

\item{beta_0}{beta for sigma^2 prior}

\item{alpha}{alpha for lambda prior}

\item{beta}{beta for lambda prior}
}
\value{
Value of posterior
}
\description{
Lambda posterior function logged
}
