% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINAL_SUPPORT_pdfs.R
\name{nlnm}
\alias{nlnm}
\alias{dnlnm}
\title{Non-local Normal Moment Distribution}
\usage{
dnlnm(x, tau2, r, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{tau2}{the tau2 parameter}

\item{r}{the r parameter}

\item{log}{logical; if \code{TRUE}, probabilities
\code{p} are given as \code{log(p)}.}
}
\value{
\code{dnlnm} gives the density of non-local
normal moment distribution
}
\description{
Non-local Normal Moment Distribution
}
