% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background_functions.R
\name{p_value_calculator_lambda}
\alias{p_value_calculator_lambda}
\title{Calculates p value by integrating at extremes}
\usage{
p_value_calculator_lambda(FUNC, v, ...)
}
\arguments{
\item{FUNC}{function to integrate over}

\item{v}{the upller value to integrate to}

\item{...}{additional inputs}
}
\value{
p value returned
}
\description{
Calculates p value by integrating at extremes
}
