% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background_functions.R
\name{poisson_unknown_rate_BFF_sequential}
\alias{poisson_unknown_rate_BFF_sequential}
\title{Calculate poisson BFF update}
\usage{
poisson_unknown_rate_BFF_sequential(
  data_new,
  N_prev,
  D_prev,
  F_prev,
  alpha_0,
  beta_0,
  alpha,
  beta
)
}
\arguments{
\item{data_new}{new data point}

\item{N_prev}{N prev}

\item{D_prev}{D prev}

\item{F_prev}{F prev}

\item{alpha_0}{rate prior alpha}

\item{beta_0}{rate prior beta}

\item{alpha}{lambda prior alpha}

\item{beta}{lambda prior beta}
}
\value{
List containing updated hyperparameters and estimates
}
\description{
Calculate poisson BFF update
}
