% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.BFF}
\alias{summary.BFF}
\title{Summarize BFF object}
\usage{
\method{summary}{BFF}(object, ...)
}
\arguments{
\item{object}{a BFF object}

\item{...}{additional arguments (unused)}
}
\value{
prints summary of a BFF object.
}
\description{
Summarize BFF object
}
\seealso{
[z_test_BFF()], [t_test_BFF()], [chi2_test_BFF()], [f_test_BFF()]
}
