% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background_functions.R
\name{threshold_p_value_detector}
\alias{threshold_p_value_detector}
\title{Detect if p value is anomlous at given threshold}
\usage{
threshold_p_value_detector(p_values, threshold_val, burn_out)
}
\arguments{
\item{p_values}{p values}

\item{threshold_val}{threshold}

\item{burn_out}{grace period}
}
\value{
anomalous p vlaues
}
\description{
Detect if p value is anomlous at given threshold
}
