% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background_functions.R
\name{threshold_p_value_detector_seq}
\alias{threshold_p_value_detector_seq}
\title{Detect if p value is anomlous at given threshold sequential version}
\usage{
threshold_p_value_detector_seq(
  i,
  p_value,
  current_anom,
  threshold_val,
  burn_out
)
}
\arguments{
\item{i}{current data number}

\item{p_value}{p value}

\item{current_anom}{current list of anomalous values}

\item{threshold_val}{threshold}

\item{burn_out}{grace period}
}
\value{
anomalous p vlaues
}
\description{
Detect if p value is anomlous at given threshold sequential version
}
