% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINAL_FUNCTIONS_chi2_test.R
\name{chi2_test_BFF}
\alias{chi2_test_BFF}
\title{chi2_test_BFF}
\usage{
chi2_test_BFF(chi2_stat, n, df, LRT = FALSE, r = NULL, omega = NULL)
}
\arguments{
\item{chi2_stat}{chi-square statistic}

\item{n}{sample size (if one sample test)}

\item{df}{degrees of freedom}

\item{LRT}{should LRT be performed? Default is FALSE}

\item{r}{r value}

\item{omega}{omega values (can be a single entry or a vector of values)}
}
\value{
Returns Bayes factor function results
 \tabular{ll}{
   \code{BFF} \tab The object containing the log_bf (log bayes factor values) and corresponding omega values \cr
   \tab \cr
   \code{log_bf} \tab maximized bayes factor\cr
   \tab \cr
   \code{omega_set} \tab omega value corresponding to maximized bayes factor\cr
   \tab \cr
   \code{omega_set} \tab was an omega value provided?\cr
   \tab \cr
   \code{alternative} \tab user provided alternative \cr
   \tab \cr
   \code{f} \tab final t value if maximized, or input r value if provided \cr
   \tab \cr
}
}
\description{
chi2_test_BFF constructs BFFs based on the t test. BFFs depend on hyperparameters r and tau^2 which determine the shape and scale of the prior distributions which define the alternative hypotheses.
By setting r > 1, we use higher-order moments for replicated studies. Fractional moments are set with r > 1 and r not an integer.
All results are on the log scale.
}
\examples{
chi2BFF = chi2_test_BFF(chi2_stat = 7.5, n = 25, df = 23)
chi2BFF
plot(chi2BFF)

}
