% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINAL_FUNCTIONS_chi2_test.R
\name{chi2_test_BFF}
\alias{chi2_test_BFF}
\title{chi2_test_BFF}
\usage{
chi2_test_BFF(chi2_stat, n, df, LRT = FALSE, r = NULL, omega = NULL)
}
\arguments{
\item{chi2_stat}{chi-square statistic}

\item{n}{sample size (if one sample test)}

\item{df}{degrees of freedom}

\item{LRT}{should LRT be performed? Default is FALSE}

\item{r}{r value}

\item{one_sample}{is test one sided? Default is FALSE}

\item{tau2}{tau2 values (can be a single entry or a vector of values)}
}
\value{
Returns Bayes factor function results
 \tabular{ll}{
   \code{BFF} \tab The log of the Bayes Factor Function values \cr
   \tab \cr
   \code{effect_size} \tab Effect sizes tested (seq(0, 1, by = 0.01)) \cr
   \tab \cr
   \code{BFF_max_RMSE} \tab Maximum BFF value \cr
   \tab \cr
   \code{max_RMSE} \tab Effect size that maximizes BFF\cr
   \tab \cr
   \code{omega} \tab omega values tested, can be a single number or vector\cr
}
}
\description{
chi2_test_BFF constructs BFFs based on the t test. BFFs depend on hyperparameters r and tau^2 which determine the shape and scale of the prior distributions which define the alternative hypotheses.
By setting r > 1, we use higher-order moments for replicated studies. Fractional moments are set with r > 1 and r not an integer.
All results are on the log scale.
}
\examples{
chi2BFF = chi2_test_BFF(chi2_stat = 2.5, n = 50, df = 48)
chi2_test_BFF(chi2_stat = 2.5, n = 50, tau2 = c(0.5, 0.8))
chi2BFF$BFF_max_RMSE   # maximum BFF omega
chi2BFF$max_RMSE       # effect size which maximizes the BFF value

}
