% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINAL_FUNCTIONS_f_test.R
\name{f_test_BFF}
\alias{f_test_BFF}
\title{f_test_BFF}
\usage{
f_test_BFF(f_stat, n, df1, df2, r = NULL, omega = NULL)
}
\arguments{
\item{f_stat}{T statistic}

\item{n}{sample size (if one sample test)}

\item{df1}{sample size of group one for two sample test.}

\item{df2}{sample size of group two for two sample test}

\item{r}{r value}

\item{omega}{omega values (can be a single entry or a vector of values)}
}
\value{
Returns an S3 object with Bayes Factor function results.
 \tabular{ll}{
   \code{BFF} \tab the object containing the log_bf (log bayes factor values) and corresponding omega values \cr
   \tab \cr
   \code{input} \tab the object containing the input values \cr
   \tab \cr
   \code{log_bf} \tab maximized bayes factor\cr
   \tab \cr
   \code{omega} \tab corresponding omega value for maximized bayes factor\cr
   \tab \cr
   \code{omega_set} \tab was an omega value provided?\cr
   \tab \cr
   \code{r} \tab r value (default is 1 if not provided by user) \cr
   \tab \cr
   \code{test_type} \tab type of BFF test\cr
   \tab \cr
   \code{generic_test} \tab FALSE \cr
   \tab \cr
}
}
\description{
f_test_BFF constructs BFFs based on the t test. BFFs depend on hyperparameters r and tau^2 which determine the shape and scale of the prior distributions which define the alternative hypotheses.
By setting r > 1, we use higher-order moments for replicated studies. Fractional moments are set with r > 1 and r not an integer.
All results are on the log scale.
}
\examples{
fBFF = f_test_BFF(f_stat = 2.5, n = 50, df1 = 25, df2 = 48)
fBFF
plot(fBFF)

}
