% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.BFF}
\alias{plot.BFF}
\title{Plot Bayes Factor Function}
\usage{
\method{plot}{BFF}(x, xlab = NULL, ylab = NULL, main = NULL, ...)
}
\arguments{
\item{x}{a BFF object}
}
\value{
either a ggplot2 object if \code{plot = TRUE} or a data.frame
with a Bayes factor function if \code{plot = FALSE}
}
\description{
Creates a Bayes factor function plot
of an BFF object. The BFF object needs to be specified
with \code{omega = NULL}.
}
\seealso{
[z_test_BFF()], [t_test_BFF()], [chi2_test_BFF()], [f_test_BFF()]
}
