% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{posterior_plot}
\alias{posterior_plot}
\title{Plot Prior and Posterior Distribution}
\usage{
posterior_plot(x, prior = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{a BFF object}

\item{prior}{whether prior distribution should
be added to the figure}

\item{plot}{whether plot should be generated.
Default to \code{TRUE}. If \code{FALSE} a data
frame with the support, prior ordinates, and
posterior ordinates is returned instead.}

\item{...}{additional arguments}
}
\description{
Creates a prior and posterior plot
of an BFF object. If no specific omega was set
when fitting the model, the omega resulting in
maximum BF against the null hypothesis is selected.
}
