% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINAL_FUNCTIONS_regression_test.R
\name{regression_test_BFF}
\alias{regression_test_BFF}
\title{regression_test_BFF}
\usage{
regression_test_BFF(
  t_stat,
  n,
  k,
  alternative = "two.sided",
  r = NULL,
  omega = NULL
)
}
\arguments{
\item{t_stat}{T statistic}

\item{n}{sample size (if one sample test)}

\item{k}{number of predictors}

\item{r}{r value}

\item{tau2}{tau2 values (can be a single entry or a vector of values)}
}
\value{
Returns Bayes factor function results
 \tabular{ll}{
   \code{BFF} \tab The log of the Bayes Factor Function values \cr
   \tab \cr
   \code{effect_size} \tab Effect sizes tested (seq(0, 1, by = 0.01)) \cr
   \tab \cr
   \code{BFF_max_RMSE} \tab Maximum BFF value \cr
   \tab \cr
   \code{max_RMSE} \tab Effect size that maximizes BFF\cr
   \tab \cr
   \code{omega} \tab omega values tested, can be a single number or vector\cr
}
}
\description{
regression_test_BFF constructs BFFs based on the t test. BFFs depend on hyperparameters r and tau^2 which determine the shape and scale of the prior distributions which define the alternative hypotheses.
By setting r > 1, we use higher-order moments for replicated studies. Fractional moments are set with r > 1 and r not an integer.
All results are on the log scale.
}
\examples{
regBFF = regression_test_BFF(t_stat = 2.5, n = 50, k = 3)
regBFF
plot(regBFF)

}
