% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINAL_FUNCTIONS_z_test.R
\name{z_test_BFF}
\alias{z_test_BFF}
\title{z_test_BFF}
\usage{
z_test_BFF(
  z_stat,
  n = NULL,
  one_sample = FALSE,
  alternative = "two.sided",
  n1 = NULL,
  n2 = NULL,
  r = NULL,
  omega = NULL
)
}
\arguments{
\item{z_stat}{Z statistic}

\item{n}{sample size (if one sample test)}

\item{one_sample}{is test one sided? Default is FALSE}

\item{alternative}{the alternative. options are "two.sided" or "less" or "greater"}

\item{n1}{sample size of group one for two sample test. Must be provided if one_sample = FALSE}

\item{n2}{sample size of group two for two sample test. Must be provided if one_sample = FALSE}

\item{r}{r value}

\item{omega}{omega values (can be a single entry or a vector of values)}
}
\value{
Returns an S3 object with Bayes Factor function results.
 \tabular{ll}{
   \code{BFF} \tab the object containing the log_bf (log bayes factor values) and corresponding omega values \cr
   \tab \cr
   \code{input} \tab the object containing the input values \cr
   \tab \cr
   \code{log_bf} \tab maximized bayes factor\cr
   \tab \cr
   \code{omega} \tab corresponding omega value for maximized bayes factor\cr
   \tab \cr
   \code{one_sample} \tab is this a one sample test? \cr
   \tab \cr
   \code{alternative} \tab alternative hypothesis used in calculations \cr
   \tab \cr
   \code{omega_set} \tab was an omega value provided?\cr
   \tab \cr
   \code{r} \tab r value (default is 1 if not provided by user) \cr
   \tab \cr
   \code{alternative} \tab user provided alternative \cr
   \tab \cr
   \code{test_type} \tab type of BFF test\cr
   \tab \cr
   \code{generic_test} \tab FALSE \cr
   \tab \cr
}
}
\description{
z_test_BFF constructs BFFs based on the z test. BFFs depend on hyperparameters r and tau^2 which determine the shape and scale of the prior distributions which define the alternative hypotheses.
By setting r > 1, we use higher-order moments for replicated studies. Fractional moments are set with r > 1 and r not an integer.
All results are on the log scale.
}
\examples{
zBFF = z_test_BFF(z_stat = 2.5, n = 50, one_sample = TRUE)
zBFF
plot(zBFF)
}
